+++ core/src/com/dogshitempire/cos/ui/CatSummary.java
@@ -87,10 +87,10 @@ public class CatSummary extends Table {
     public void draw(Batch batch, float alpha) {
         if(cat != null) {
             catName.setText(cat.getCatName() + " (" + cat.getGenderLetter() + ")");
-            catHealth.setText("" + cat.getStats().getHealth());
-            catHappiness.setText("" + cat.getStats().getHappiness());
-            catCleanliness.setText("" + cat.getStats().getCleanliness());
-            catHunger.setText("" + cat.getStats().getHunger());
+            catHealth.setText("" + (int)cat.getStats().getHealth());
+            catHappiness.setText("" + (int)cat.getStats().getHappiness());
+            catCleanliness.setText("" + (int)cat.getStats().getCleanliness());
+            catHunger.setText("" + (int)cat.getStats().getHunger());
         }
         else {
             catName.setText("");
